Fix the unit test files
As we moved class JIRAConnection to a separate module, we need to do the
same for the corresponding unit test.

Also, fixing some small lint issues with the existing files.
