+++ tests/lib/jira_api/test_jira_issue.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from exceptions.missing_mandatory_param_error import MissingMandatoryParamError
-from jira_api.jira_issue import create_jira_task, JiraConnection
+from jira_api.jira_issue import create_jira_task
 
 
 def test_create_jira_task_project_id_throws():
@@ -45,26 +45,6 @@ def test_create_jira_task_description_throws():
     mock_account = MagicMock()
     with pytest.raises(MissingMandatoryParamError):
         create_jira_task(mock_account, mock_issue_details)
-
-
-def test_jira_connection():
-    """
-    Tests a client to connect to a atlassian server is created and closed
-    """
-    with patch("jira.client.JIRA") as mock_jira:
-        mock_account = MagicMock()
-        mock_account.atlassian_endpoint = "https://example.atlassian.net/"
-        mock_account.username = "foo"
-        mock_account.api_token = "bar"
-
-        # Use the JiraConnection class in a with statement
-        with JiraConnection(mock_account) as conn:
-            # Check if the JIRA instance was created with the correct parameters
-            mock_jira.assert_called_once_with(
-                server="https://example.atlassian.net/", basic_auth=("foo", "bar")
-            )
-            # Check if the connection object is the mock JIRA instance
-            assert conn == mock_jira.return_value
 
 
 def test_create_jira_task_without_epic_success():
